pyproject.toml
[tool.poetry]
name = "pysonet"
version = "0.1.0"
description = "Social network by Django"
authors = ["Muhammadrizo04 <muhammadrizomuxtorov389@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.3"
django = "3.1"
djangorestframework = "3.11.1"
pillow = "7.2.0"
psycopg2 = "2.8.5"
djoser = "2.0.3"
djangorestframework-simplejwt = "4.8.0"
drf-yasg = "^1.21.7"
django-cors-headers = "<4.3.1"
django-mptt = "0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
